rm un-needed mappings

+++ lua/plugins/configs/lspconfig.lua
@@ -14,7 +14,7 @@ local _default_opts = win.default_opts
 
 win.default_opts = function(options)
    local opts = _default_opts(options)
-   opts.border = "double"
+   opts.border = "single"
    return opts
 end
